+++ restaurant/src/front_of_house.rs
@@ -1,0 +1,1 @@
+pub mod hosting;
