+++ restaurant/src/lib.rs
@@ -1,18 +1,6 @@
-mod front_of_house {
-    pub mod hosting {
-        pub fn add_to_waitlist() {}
-        
-        fn seat_at_table() {}
-    }
+mod front_of_house;
 
-    mod serving {
-        fn take_order() {}
-
-        fn serve_order() {}
-
-        fn take_payment() {}
-    }
-}
+use crate::front_of_house::hosting;
 
 mod back_of_house {
     pub enum Appetizer {
@@ -35,7 +23,6 @@ mod back_of_house {
     }
 }
 
-use crate::front_of_house::hosting;
 // relative path
 // use self::front_of_house::hosting;
